### 2025-05-09 #243 @cjpalpallatoc
VARGRAM submission
Submitting Author: (@cjpalpallatoc)
Package Name: VARGRAM
One-Line Description of Package: A Python visualization tool for genomic surveillance
Repository Link:  https://github.com/pgcbioinfo/vargram
Version submitted: 0.3.0
Version accepted: TBD

---

During a viral outbreak, the diversity of sampled sequences often needs to be quickly determined to understand the evolution of a pathogen. VARGRAM (Visual ARrays for GRaphical Analysis of Mutations) empowers researchers to quickly generate a mutation profile to compare batches of sequences against each other and against a reference set of mutations. A publication-ready profile can be generated in a couple lines of code by providing sequence files (FASTA, GFF3) or tabular data (CSV, TSV, Pandas DataFrame). When sequence files are provided, VARGRAM leverages [Nextclade CLI](https://docs.nextstrain.org/projects/nextclade/en/stable/user/nextclade-cli/index.html) to perform mutation calling. We have user-friendly installation instructions and tutorials on [our documentation website](https://pgcbioinfo.github.io/vargram).

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

We hope that VARGRAM would be useful for researchers, analysts, and students in the field of molecular epidemiology/genomic surveillance. During the pandemic, we've used an early mutation profile script [to characterize emergent variants](https://www.medrxiv.org/content/10.1101/2021.03.03.21252812v2) and [potential recombinants](https://www.biorxiv.org/content/10.1101/2023.04.12.534029v3).

The closest we're aware of is [snipit](https://github.com/aineniamh/snipit). The main difference is that VARGRAM provides a visual comparison of mutation profiles between groups or within a population of samples. There are also additional features such as grouping mutations per gene, adding multiple sets of reference mutations, and other customizations. We also plan to expand the package to provide other types of visualization relevant to genomic surveillance.
We're also aware of packages like [Marsilea](https://github.com/Marsilea-viz/marsilea) that can in principle be used to make a profile, but these are more general in scope and would require more work for the user than if they used VARGRAM. Outside Python, we've seen researchers create mutation profiles with custom scripts (in R) and there are also web tools available like [Nextclade](https://clades.nextstrain.org/). VARGRAM differs by making the process substantially convenient in terms of generation and customization of the figure.

https://github.com/pyOpenSci/software-submission/issues/225
