### 2025-05-27 #243 @coatless
@cjpalpallatoc I'm happy to share we have an editor for the review!

@ctb will be the editor for this package review. I'll let him introduce himself.

> [!NOTE]
>
> We're moving the review along despite the data note given above as the sample data for examples is reachable via:
>
> https://github.com/coatless/pyopensci-eic/blob/main/i243-vargram.ipynb
>
> This will likely get picked up in _later_ individual reviews if the issue still persists.
