### 2025-06-09 5590ed7
Fixed y-axis bugs where ticks do not show when values are too low and one is added to max y-lim
