+++ src/vargram/plots/_profile_elements.py
@@ -81,7 +81,7 @@ def build_group_barplot(ax_bar, categories, heights,
     # Adjusting limits of x-axis and y-axis
     ax_bar.set_xlim(-0.5, len(categories) - 0.5)
     if max_height != 0: # Avoids UserWarning
-        ax_bar.set_ylim(0.0, max_height+1) 
+        ax_bar.set_ylim(0.0, max_height) 
     else:
         ax_bar.set_ylim(0.0, 1.0)
 
@@ -111,7 +111,6 @@ def build_group_barplot(ax_bar, categories, heights,
         ax_bar.spines['left'].set_position(('outward', 5)) 
         ax_bar.yaxis.set_tick_params(width=1.5, labelsize=y_fontsize)
         yticks = np.linspace(0, max_height, 5)
-        yticks = np.round(yticks).astype(int)
         ax_bar.set_yticks(yticks[yticks != 0])    
     return None
 
