### 2025-06-10 97e2584
Removed five-tick limit on yaxis
