### 2025-06-12 4269f65
Fixed error where parsed mutation is an empty string because sequences without mutations were not removed
