### 2025-06-12 2a40351
Changed default stack colors
