### 2025-06-12 a6bf402
Shortened Nextclade command definitions
