### 2025-06-14 a489601
Defined an internal version attribute
