+++ src/vargram/__init__.py
@@ -1,1 +1,2 @@
 from .vargram import vargram
+__version__ = "0.4.0"
