+++ pyproject.toml
@@ -4,6 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "vargram"
+dynamic = ["version"]
 description = "A Python visualization tool for genomic surveillance"
 authors = [
     {name = "C. J. Palpal-latoc"},
@@ -12,7 +13,6 @@ authors = [
 maintainers = [
     {name = "Philippine Genome Center - Core Facility for Bioinformatics", email = "bioinformatics@pgc.up.edu.ph"}
 ]
-version = "0.3.0"
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
@@ -20,7 +20,8 @@ dependencies = [
     "pandas>=2.2.2",
     "numpy>=1.26.4",
     "regex>=2024.7.24",
-    "tabulate>=0.9.0"
+    "tabulate>=0.9.0",
+    "requests>=2.32.0"
 ]
 keywords = ["virus", "mutations", "epidemiology", "visualization"]
 classifiers = [
@@ -33,6 +34,9 @@ classifiers = [
     "Programming Language :: Python :: 3.11",
     "Operating System :: OS Independent"
 ]
+
+[tool.setuptools.dynamic]
+version = {attr = "vargram.__version__"}
 
 [project.urls]
 Repository = "https://github.com/pgcbioinfo/vargram"
