### 2025-06-14 21c154a
Allowed group fontsize to be changed
