### 2025-06-14 0c22a62
Made arguments explicit for profile, aes and key methods and added group_fontsize argument for aes
