### 2025-06-14 7b8dc34
Added check for provided group names and accepted only list of lists or str for struct
