### 2025-06-16 58aa876
Removed default xlim of barplots when x-axis is numeric
