+++ src/vargram/plots/_profile_elements.py
@@ -3,6 +3,7 @@ and auxiliary functions."""
 
 import matplotlib.colors as mc
 import numpy as np
+import pandas as pd
 
 
 def check_xticks_overlap(xticks):
@@ -79,7 +80,18 @@ def build_group_barplot(ax_bar, categories, heights,
     ax_bar.spines["bottom"].set_visible(False)
 
     # Adjusting limits of x-axis and y-axis
-    ax_bar.set_xlim(-0.5, len(categories) - 0.5)
+    is_numeric_dtype = pd.api.types.is_numeric_dtype(categories)
+    #if is_numeric_dtype:
+    #    xticks = ax_bar.get_xticks()
+    #    xmin = min(categories)
+    #    if xmin not in xticks:
+    #        new_xticks = list(xticks) + [xmin]
+    #        new_xticks = sorted(set(new_xticks))
+    #        ax_bar.set_xticks(new_xticks)
+        #xmax = max(categories)
+        #ax_bar.set_xlim(xmin-0.5,xmax)
+    if not is_numeric_dtype:
+        ax_bar.set_xlim(-0.5, len(categories) - 0.5)
     if max_height != 0: # Avoids UserWarning
         ax_bar.set_ylim(0.0, max_height) 
     else:
