### 2025-06-16 d407466
Added figsize argument to aes
