### 2025-06-16 a5a9cde
Fixed error due to division by zero when entire stack sum is zero
