### 2025-06-16 c1c357d
Closed figure when saving tabulated file
