### 2025-06-16 abf366a
Fixed default legend title error by separating the stack and group title checks for user input
