+++ docs/mutation_profile.md
@@ -15,6 +15,13 @@ A mutation profile such as the one above is an intuitive figure that shows you w
 
 ## Input
 
+### Importing VARGRAM
+
+To import the VARGRAM class, simply run:
+```py
+from vargram import vargram
+```
+
 ### Providing sequence files 
 There are two main sets of input that you can provide. The first is the sequence files which include (1) the FASTA file of sequences or a directory of FASTA files for multiple batches, (2) the reference FASTA, and (3) a genome annotation file following [the GFF3 format](https://docs.nextstrain.org/projects/nextclade/en/stable/user/input-files/03-genome-annotation.html). An example annotation file is shown below:
 ```
@@ -46,14 +53,14 @@ Instead of providing the reference FASTA, you may also specify the name or short
 When sequence files are provided, VARGRAM will run Nextclade CLI and capture the analysis file so make sure that Nextclade is [installed](install_nextclade.md). These files can be provided to VARGRAM through the `vargram` class:
 === "Local reference"
     ```py
-    vg = vargram(seq='path/to/covid_samples/', # Provide FASTA of sequences or directory containing FASTA files
-                ref='path/to/covid_reference.fa', # Provide reference sequence
-                gene='path/to/covid_annotation.gff') # Provide genome annotation
+    vg = vargram(seq='path/to/<samples/>', # Provide FASTA of sequences or directory containing FASTA files
+                ref='path/to/<reference.fa>', # Provide reference sequence
+                gene='path/to/<annotation.gff>') # Provide genome annotation
     ```
 === "Online Nextclade reference"
     ```py
-    vg = vargram(seq='path/to/covid_samples/', # Provide sample sequences
-                ref='sars-cov-2') # Provide reference name/shortcut
+    vg = vargram(seq='path/to/<samples/>', # Provide sample sequences
+                ref='<reference_name>') # Provide reference name/shortcut
     ```
 
 !!! warning "Nextclade dataset name"
@@ -73,13 +80,13 @@ When sequence files are provided, VARGRAM will run Nextclade CLI and capture the
 
 You may already have a Nextclade analysis file obtained through Nextclade CLI or Nextclade Web. In this case, simply provide its path:
 ```py
-vg = vargram(data='path/to/nextclade_analysis.csv')
+vg = vargram(data='path/to/<analysis.csv>')
 ```
 If you want to order the genes in the profile, you must also provide the annotation file using the same keyword argument `gene`.
 
 The VARGRAM data output can also be provided as an input but you must specify its format:
 ```py
-vg = vargram(data='path/to/vargram_analysis.csv',
+vg = vargram(data='path/to/<vargram_output.csv>',
              format='vargram')
 ```
 
@@ -87,44 +94,118 @@ vg = vargram(data='path/to/vargram_analysis.csv',
 
     Nextclade CLI and Nextclade Web yield identical results when the datasets they use are also the same. If you get different mutation profiles, try updating your dataset. See this [discussion on Nexstrain](https://discussion.nextstrain.org/t/nextclade-cli-clade-calls-different-from-nextclade-web-calls/898) for more info. 
 
+
+### Downloading test data
+
+Test data from VARGRAM, which can be downloaded from the [VARGRAM repo](https://github.com/pgcbioinfo/vargram/tree/main/tests/test_data), and even datasets from Nextclade can be downloaded. First, import the data download modules:
+```py
+from vargram.data import example, nextclade
+```
+The `example` module downloads VARGRAM test data and the `nextclade` module is for Nextclade datasets. Using the latter module requires Nextclade CLI to be installed.
+
+To download VARGRAM test data, run
+```py
[…]
