+++ docs/basics.md
@@ -2,11 +2,20 @@
 
 ## Example
 
-If you wish to get started immediately, [install VARGRAM](install_vargram.md) (`pip install vargram`) and [download the test data from the VARGRAM repo](https://github.com/pgcbioinfo/vargram/tree/main/tests/test_data/). Then, run either of the following code snippets on a Jupyter notebook or as a Python script in the directory containing those test files (e.g. inside `test_data/` if you downloaded the entire folder):
+If you wish to get started immediately, [install VARGRAM](install_vargram.md) (`pip install vargram`). Then, run either of the following code snippets on a Jupyter notebook or as a Python script. The code below downloads [test data from the VARGRAM repo](https://github.com/pgcbioinfo/vargram/tree/main/tests/test_data) and does not require [Nextclade CLI](https://pgcbioinfo.github.io/vargram/install_nextclade/).
 === "Simple script for bare figure"
     The following is the simplest way to generate a mutation profile with default settings:
     ```py
-    vg = vargram(data='analysis/omicron_analysis_cli.tsv') # Provide data
+    # Import main VARGRAM module and module to download external data
+    from vargram import vargram
+    from vargram.data import example
+
+    # Download test data into directory called test_data 
+    # Directory is created if it does not yet exist
+    example.get('test_data') 
+
+    # Generate the mutation profile
+    vg = vargram(data='test_data/analysis/omicron_analysis_cli.tsv') # Provide data
     vg.profile() # Tell VARGRAM you want to create a mutation profile
     vg.show() # Show the figure
     vg.save("default_profile.png", dpi=300) # Save the figure
@@ -14,9 +23,9 @@ If you wish to get started immediately, [install VARGRAM](install_vargram.md) (`
     This will produce the following bare figure:
     ![plot](assets/images/default_profile.png)
 
-    The profile consists of several barplots stacked together; there is one barplot per mutated gene (gene name is shown in rectangular box just above the barplot). The x-axis of each barplot shows the amino acid mutations (substitution, deletion, or insertion) in the corresponding gene and the y-axis shows either the raw number ("counts") or the proportion ("weights") of sequences with those mutations. The bars are colored per sequence batch.
+    The profile consists of several barplots stacked together; there is one barplot per mutated gene (gene name is shown in rectangular box just above the barplot). The x-axis of each barplot shows the amino acid mutations (substitutions, deletions, and insertions) in the corresponding gene and the y-axis shows either the raw number ("counts") or the proportion ("weights") of sequences with those mutations. The bars are colored per sequence batch.
 
-    By default, the batch labels in the legend are the filenames of the provided FASTA files or `my_batch` if a Nextclade analysis CSV file is provided instead. If the gene name does not fit within the rectangular box, the gene name is replaced with a number and a legend is automatically created for it.
[…]
