### 2025-06-16 #243 @cjpalpallatoc
Hi @coatless and @ctb !

I released a new version of VARGRAM ([v0.4.0](https://github.com/pgcbioinfo/vargram/releases/tag/v0.4.0)). Sorry this took a while. I've now added modules to download test data as well as datasets from Nextclade:
```python
from vargram.data import example, nextclade

example.get() # Download VARGRAM test data
nextclade.get('sars-cov-2') # Download SARS-CoV-2 dataset from Nextclade
```

All data files are now in just one directory: [`test_data/`](https://github.com/pgcbioinfo/vargram/tree/main/tests/test_data). (Some of them were previously in the `docs/` directory.) I also want to clarify that `nextclade_analysis.csv` and `covid_samples/*` were meant to be generic names. To avoid confusion, I now added angle brackets for this where necessary. The examples in the [tutorial](https://pgcbioinfo.github.io/vargram/mutation_profile/) also now use actual test data that the user can copy and run.

Hope this resolves the issue!
