### 2025-07-08 #243 @ctb
hi @cjpalpallatoc apologies for delay - starting to look for reviewers :).  I got the package installed and ran the example script, so things are good there!
